FastbootCmds: reset stored rollback indices when unlocking bootloader
When unlocking the device also reset the stored rollback indices to
make it easier to relock bootloaders.

Usecase:
  1) Stock with SPL 2025-05-05 is installed
  2) Bootloader is unlocked
  3) Custom ROM with SPL set to 2024-10-05 is flashed
  4) Bootloader is relocked

Without resetting of SPL, this would not work and the devices
would need to be unlocked forever.

If the custom rom would then bump the SPL to something higher
it would be possible, that the device can not properly return
to stock, as the bootloader does not check for what kind of OS
is installed when bumping the rollback indices.

Note: for security cases, where we must not allow downgrading
to a certain version, we can still implement this when needed.

Issue: FP4-A14#496
Change-Id: Icbad2655d629ab39f8aec7a3e2f0155fa0497c1e
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>
[luca.weiss@fairphone.com: Replace references of SPL with rollback
index]